Let USE flag 'feroceon' control whether write-magic patch should be applied
If you use the regular shutdown or halt command on the Excito B3 it will
become unresponsive and the purple led will stay lit. Instead you must write
a special 'magic' value to its flash memory and reboot, at which point the
u-boot environment will turn off the front led and wait for the button on the
back to be pressed before booting the system again. The write-magic patch adds
this logic to the regular shutdown command, so the Excito B3 will seemingly
respond to the shutdown command like any other regular machine.