Change version ID on sysvinit from 'rolling' to its actual version
Do note the high revision number which is to avoid conflicts with
updates to the regular Gentoo package. This version specifically
targets the Excito B3 platform, allowing you to use the regular
shutdown commands instead of the complex procedure of writing to
flash memory and then run 'reboot'.

To install it you must accept keyword "**" and you are advised to
mask >=sys-apps/sysvinit-2.89 as any other version will cause the
Excito B3 to ill-behave when issuing 'halt' or 'poweroff' command
(or their 'shutdown' command equivalents).

This package includes a hardware check and will not install on
any system not running the exact same SoC as the Excito B3. While
a USE flag is provided to circumvent that hardware check it is
quite useless to do so as this will result in the exact same
binaries being installed as with the regular Gentoo release
sysvinit-2.88-r7 (i.e. the additional Excito B3 hardware specific
patches will then be skipped).