LogitechMediaServer is not a perl module
Replaced the perl-module eclass inclusion with explicit perl_features_ithreads
USE flag. This may still create a confusing situation since the package does
not do anything with this USE flag other than call die() if it is not enabled.
A mayor annoyance is that we can't actually check whether the USE flag was
passed on as the required global PERL_FEATURES use-expand variable or as a
package individual USE flag, so users may still unintentionally break their
system if they misinterpret the instructions.